Move browsing components into fragment to support tabs
